backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1

# HTTP client
httpx==0.28.1
requests==2.32.0

# Utilities
python-dateutil==2.9.0
websockets==14.1
